backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.2
numpy==1.26.0
pandas==2.1.1
scipy==1.11.3
yfinance==0.2.28
python-dotenv==1.0.0

# Quantum optimization packages (optional)
qiskit==1.2.4
qiskit-aer==0.17.0
qiskit-optimization==0.7.0
qiskit-algorithms==0.3.1